tests: simplify checks for some expected variables values in Makefiles
Do so by using our custom 'is' auxiliary script rather than grepping
the output from make.  This is more natural, more robust, and often
shorter to write.

Unfortunately, we can't do that in all cases: sometimes we really need
to match the content of a variable against a regular expressions, and
we can't know nor are interested in its exact value.

This is basically a follow-up on commit v1.11-1830-g96401cb of
2012-02-08 (tests: better way to compare lists in Makefile rules).

* t/subst-no-trailing-empty-line.sh: Adjust.
* t/pluseq10.sh: Likewise.
* t/check5.sh: Likewise, and enhance a little while at it.
* t/check7.sh: Likewise.
* t/exeext.sh: Likewise.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>