distcheck: don't allow overriding of --prefix and --srcdir by the user
Not through AM_DISTCHECK_FLAGS, nor through DISTCHECK_FLAGS.  Apparently,
some packages got in the habit of relaying all the options passed to the
original ./configure invocation through to the configure invocations
in "make distcheck".  This was causing problems, because it also passed
through the original --srcdir and --prefix options.

Fixes: expose bug#14991 (relates to 'distcheck')

* lib/am/distdir.am (distcheck): Pass the hard-coded --srcdir and
--prefix options *after* both the developer-defined options in
$(AM_DISTCHECK_FLAGS) and the user-defined options in $(DISTCHECK_FLAGS).
* t/list-of-tests.mk (XFAIL_TESTS): Remove the now-passing test
'distcheck-no-destdist-or-srcdir-override.sh'.
* doc/automake.texi (Checking the Distribution): Update.
* NEWS: Likewise.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>