build: Inline perl prototypes in sources
Prototypes allows us to avoid using the '&foo' invocation form when
invoking a subroutine before its definition.  Previously those
prototypes were generated to prevent them from falling out-of-sync
with actual definitions.  Now we provide a 'check-perl-protos' lint
script to ensure that this is not the case.  This has the same
benefits as generating prototypes while simplifying the
bootstrap/build process.

* bin/gen-perl-protos: Remove.
* bin/Makefile.inc: Adapt.
* bootstrap: Likewise.
* bin/aclocal.in: Inline prototypes.
* bin/automake.in: Likewise.
* maintainer/check-perl-protos: New lint script.
* maintainer/syntax-checks.mk (sc_perl_protos): New target.
(syntax_check_rules): Add it.