check: Make 'check-TESTS' target depend on test dependencies
The 'check' and 'recheck' targets are already depending on the progams
and scripts required to run the test suite.  The 'check-TESTS' target
while not in the public API might be used by uninformed users, so it
makes sense to add the same dependency to it.  This fixes bug#27186.

* lib/am/check.am [!%?SERIAL_TESTS%] (check-TESTS): Add dependency to
'%CHECK_DEPS%'.
* THANKS: Update.