mdate-sh: Ensure reproducible time output
This change fixes automake bug#20314.

'mdate-sh' pretty-prints the modification time of a file.  But it's
output can vary depending on the timezone of the caller. Someone in
timezone GMT-12 will get a different result (day) than someone in
timezone GMT+12.  As this output is also used to create/update stamp
files, which influence the further build process, the build result can
vary.

* lib/mdate-sh: Set 'TZ' to UTC which ensures reproducible output.
* NEWS: Announce bug fix.

Copyright-paperwork-exempt: yes