Move $gen_copyright
We move the variable to the bottom of the file where it is used because we
have no real reason to declare it where it was.  The added comment has also
been removed since it doesn't make sense and isn't useful for the
comprehension of this variable.  The variable will stay in this file unless we
have a real reason to move it out of this file.