Add test for Automake::File
This tests the make_paragraphs method from the Automake::File library.  This
uses the newly added .plt file extension as we use the Test::Simple library.

- t/pm/File.pl: Add it.

FIXME: Their is a bug with the test that says:

Filehandle STDOUT reopened as $fh only for input at
/usr/share/perl/5.26/Test2/IPC/Driver/Files.pm line 144 during global
destruction.

It only happens when Automake modules are included.  I suspect this comes from
the Automake::General module but this needs to be investigated.