t/python-virtualenv.sh: Skip when versions don’t match
On some operating systems ‘python’ is Python 2.x but ‘virtualenv -ppython’
will create a virtualenv that uses Python 3.x.  This is a bug, but it’s
not *automake’s* bug, and should not cause t/python-virtualenv.sh to fail.
Skip the test, instead of failing it, when the inner=outer version check
fails.

(This also has nothing to do with the main goal of this patchset, it just
annoyed me while I was testing.)

* t/python-virtualenv.sh: Skip test, rather than failing it, when
  $py_version_pre != $py_version_post.