tests: recompute dependencies when lists-of-tests.mk changes.
This change fixes https://bugs.gnu.org/44458
and updates https://bugs.gnu.org/11347.

* t/local.mk ($(srcdir)/%D/testsuite-part.am): restore
dependency on '%D/list-of-tests.mk' (i.e., t/list-of-tests.mk),
partially reverting the change of 26 Apr 2012 for bug#11347.
Otherwise, new tests that have dependencies will not cause an
update of testsuite-part.am, leading to strange failures
(bug#44458). The original problem being fixed in #11347 was
unnecessary rebuilding when modifying tests; that should not be
affected here, but when new tests are added, it seems reasonable,
as well as necessary, to ensure dependencies are updated.