tests: add coverage for nobase_ and dist_ prefixes
From automake bug https://bugs.gnu.org/14245.

We have tests that check nobase_ and nodist_ interaction, but not
nobase_ and dist_.  We had a report against automake 1.12 that this
combo did not work.  It seems to work now, but lets include a test
to make sure we don't regress.

Based on Daiki Ueno's example, but fixed up to pass against master.

* t/list-of-tests.mk: Add t/nobase-dist.sh.
* t/nobase-dist.sh: New test.