python: fix exit status handling with uninstall
The st variable is set at the top of this shell script, and then here
is a pipeline where it tries to update it in the subshell.  But since
setting variables in a subshell doesn't propagate back up, it doesn't
actually work.  Have the final subshell in the pipeline manage its own
exit status and exit with that so that the final status of the pipeline
is the right value.

* lib/am/python.am: Fix final subshell exit status passing.