tests: depcomp: ensure make_ok() fails when run_make fails
While running automake tests with -std-gnu=c99, the compiler report
errors which lead to make to fail. Yet, these failures are ignored
during the tests, which considers them to be successful as stderror is
check for one specific pattern.

If make fails, investigation should be made to discover why, whatever
the reason for the failure is.

* t/ax/depcomp.sh: Make make_ok fail when make fails.