depend: trim spurious leading tab
These vars are replaced with a list of remove commands that are joined
with a \n and each line always has a leading \t inserted.  That means
the literal tab here before the var leads to 2 tabs included in the
output.  While not functionally a problem, it can be a bit confusing
when reading the output as it implies something is amiss.  Trimming
the tab also aligns with the other .am files which don't include it.

* lib/am/depend.am: Delete tab before %DISTRMS%.