Adding the "bootstrap" instructions
It may be obvious, but I didn't understand how to create the config file from the configure.ac. I didn't see the command to run "./bootstrap," but it ended up working for me.  We don't have to add the instructions here, but maybe this should be mentioned in the INSTALL directions?