automake: algol68: rename A68 to A68C and AM_PROG_A68 to AM_PROG_A68C.
From https://bugs.gnu.org/78033 (automake-patches).

While upstreaming the Algol 68 autoconf support it was pointed
out that the shell variable A68 should be renamed to A68C and
that the macro AC_PROG_A68 should be renamed to AC_PROG_A68C.
This is to be more consistent to the support for other languages
like Go. This patch updates the Automake side.

The Algol 68 code snippets are also updated to use modern
stropping, which is the stropping regime now used by default by ga68.

* bin/automake.in (a68): change config_vars, libtool_tags et al.
* lib/Automake/Variable.pm: likewise rename.
* doc/automake.texi: likewise rename.
* t/a68-demo.sh: likewise modify.